db: Cleanup the server using the sqlalchemy ORM facility
We already include the entirety of sqlalchemy, so there is little point in not
using the ORM framework to avoid mapping things manually. We can also express
the schema in pure python, with input validation wherever required. And
finally we can grab the DATABASE_URL information from the environment and test
against an in-memory sqlite3 database if not in production.